@@ -1,22 +1,22 @@
-ruby '2.2.0'
+ruby '2.5.1'
 source 'https://rubygems.org'
 
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '4.1.2'
+gem 'rails', '~> 5.1'
 # Use mysql as the database for Active Record
-gem 'mysql2'
+gem 'mysql2', '< 0.5'
 gem 'rails_12factor'
 gem 'figaro'
 
 group :assets do
-  gem 'sass-rails', '~> 4.0.3'
-  gem 'coffee-rails', '~> 4.0.0'
-  gem 'uglifier', '>= 1.3.0'
+  gem 'sass-rails', '~> 5.0'
+  gem 'coffee-rails', '~> 4.2'
+  gem 'uglifier', '~> 4.1'
   #gem 'turbo-sprockets-rails3' # faster asset compilation
 end
 
 
-gem 'less-rails'
+gem 'less-rails', '~> 2.8'
 # See https://github.com/sstephenson/execjs#readme for more supported runtimes
 gem 'therubyracer',  platforms: :ruby
 gem 'twitter-bootstrap-rails'
@@ -28,10 +28,10 @@ gem 'turbolinks'
 # Build JSON APIs with ease. Read more: https://github.com/rails/jbuilder
 gem 'jbuilder', '~> 2.0'
 # bundle exec rake doc:rails generates the API under doc/api.
-gem 'sdoc', '~> 0.4.0',          group: :doc
+gem 'sdoc', '~> 0.4',          group: :doc
 
 # Spring speeds up development by keeping your application running in the background. Read more: https://github.com/rails/spring
-gem 'spring', '~> 1.3.5' ,      group: :development
+gem 'spring', '~> 1.3' ,      group: :development
 
 # Use ActiveModel has_secure_password
 # gem 'bcrypt', '~> 3.1.7'
@@ -46,10 +46,20 @@ group :development, :test do
   gem 'pry'
   gem 'better_errors'
   gem "binding_of_caller"
+  gem 'byebug'
   gem 'capistrano', '~> 3.3.0', require: false
   gem 'capistrano-rails',   '~> 1.1', require: false
   gem 'capistrano-bundler', '~> 1.1', require: false
   gem 'capistrano-rbenv', '~> 2.0', require: false
+  gem 'rspec-rails', '~> 3.7'
+  gem 'factory_bot_rails', '~> 4.10'
+end
+
+group :test do
+  gem 'capybara-webkit', '~> 1.15'
+  gem 'database_cleaner'
+  gem 'simplecov', '~> 0.16'
+  gem 'launchy'
 end
 
 # logging
@@ -65,13 +75,12 @@ gem 'rmagick'
 gem 'aws-s3'
 
 # translation
-gem 'globalize', '~> 4.0.1'
-gem 'globalize-versioning', '~> 0.1.0'
+gem 'globalize', '~> 5.0'
 
 gem 'metamagic'
 
 # Manage authentication
-gem 'devise'
+gem 'devise', '~> 4.4'
 
 # Pagination gem
 gem 'kaminari'
@@ -88,6 +97,6 @@ gem 'nprogress-rails'
 gem 'ckeditor_rails'
 
 # for frindly urls
-gem 'friendly_id', '~> 5.0.0'
+gem 'friendly_id', '~> 5.0'
 
 gem "select2-rails"
